NMEASentence/__getattr__: Graceful fail
** Return None and handle the error without raising an error. Technically
raising an error after catching another is bad as it still creates a stack
trace (causing a hard failure) and it demolishes the original stack. Here the
return None is handled and a message is logged in debug format. The other way
to resolve this on a hard failure is to reraise the origin after logging just call
raise to preserve the stack.

** Altered the test that utilized this code.